few size optimizations and that's it xd

+++ src/player.h
@@ -24,7 +24,7 @@
 #define PLAYER_H
 
 #ifdef HAVE_CONFIG_H
-#  include "config.h"
+# include "config.h"
 #endif
 
 #include <string>
@@ -43,12 +43,12 @@
 #include "sidlib_features.h"
 
 #ifdef HAVE_TSID
-#  if HAVE_TSID > 1
-#    include <tsid2/tsid2.h>
-#    define TSID TSID2
-#  else
-#    include <tsid/tsid.h>
-#  endif
+# if HAVE_TSID > 1
+#   include <tsid2/tsid2.h>
+#   define TSID TSID2
+# else
+#   include <tsid/tsid.h>
+# endif
 #endif
 
 typedef enum { black, red, green, yellow, blue, magenta, cyan, white }
@@ -95,67 +95,67 @@ void displayError (const char *arg0, unsigned int num);
 class ConsolePlayer {
 private:
 #ifdef HAVE_SIDPLAYFP_BUILDERS_RESIDFP_H
-    static const char  RESIDFP_ID[];
+    static const char RESIDFP_ID[];
 #endif
 #ifdef HAVE_SIDPLAYFP_BUILDERS_RESID_H
-    static const char  RESID_ID[];
+    static const char RESID_ID[];
 #endif
 #ifdef HAVE_SIDPLAYFP_BUILDERS_HARDSID_H
-    static const char  HARDSID_ID[];
+    static const char HARDSID_ID[];
 #endif
 #ifdef HAVE_SIDPLAYFP_BUILDERS_EXSID_H
-    static const char  EXSID_ID[];
+    static const char EXSID_ID[];
 #endif
 #ifdef HAVE_TSID
-    TSID               m_tsid;
+    TSID              m_tsid;
 #endif
 
-    const char* const  m_name;
-    sidplayfp          m_engine;
-    SidConfig          m_engCfg;
-    SidTune            m_tune;
-    player_state_t     m_state;
-    const char*        m_outfile;
-    std::string        m_filename;
+    const char* const m_name;
+    sidplayfp         m_engine;
+    SidConfig         m_engCfg;
+    SidTune           m_tune;
+    player_state_t    m_state;
+    const char*       m_outfile;
+    std::string       m_filename;
 
-    IniConfig          m_iniCfg;
-    SidDatabase        m_database;
+    IniConfig         m_iniCfg;
+    SidDatabase       m_database;
 
-    uint8_t            m_registers[3][32];
-    uint16_t*          m_freqTable;
+    uint8_t           m_registers[3][32];
+    uint16_t*         m_freqTable;
 
     // Display parameters
-    uint_least8_t      m_quietLevel;
-    uint_least8_t      m_verboseLevel;
+    uint_least8_t m_quietLevel;
+    uint_least8_t m_verboseLevel;
 
-    bool               m_cpudebug;
+    bool m_cpudebug;
 
-    bool               newSonglengthDB;
+    bool newSonglengthDB;
 
     bool vMute[9];
 
-    int  m_channels;
-    int  m_precision;
+    uint8_t m_channels;
+    uint8_t m_precision;
 
     struct m_filter_t {
         // Filter parameter for reSID
-        double         bias;
+        double bias;
         // Filter parameters for reSIDfp
-        double         filterCurve6581;
-        double         filterCurve8580;
+        double filterCurve6581;
+        double filterCurve8580;
 
-        bool           enabled;
+        bool   enabled;
     } m_filter;
 
     struct m_driver_t {
-        OUTPUTS        output;   // Selected output type
-        SIDEMUS        sid;      // Sid emulation
-        bool           file;     // File based driver
-        bool           info;     // File metadata
-        AudioConfig    cfg;
-        IAudio*        selected; // Selected Output Driver
-        IAudio*        device;   // HW/File Driver
-        Audio_Null     null;     // Used for everything
+        OUTPUTS     output;   // Selected output type
+        SIDEMUS     sid;      // SID emulation
+        bool        file;     // File based driver
+        bool        info;     // File metadata
+        AudioConfig cfg;
+        IAudio*     selected; // Selected output driver
+        IAudio*     device;   // HW/File Driver
+        Audio_Null  null;     // anything else
     } m_driver;
 
     struct m_timer_t { // secs
@@ -173,7 +173,7 @@ private:
         uint_least16_t songs;
         bool           loop;
         bool           single;
-	int            query;
+	    int            query;
     } m_track;
 
     struct m_speed_t {
@@ -183,22 +183,22 @@ private:
 
 private:
     // Console
-    void consoleColour  (player_colour_t colour, bool bold);
-    void consoleTable   (player_table_t table);
-    void consoleRestore (void);
+    void consoleColour (player_colour_t colour, bool bold);
+    void consoleTable  (player_table_t table);
+    void consoleRestore(void);
 
     // Command line args
-    void displayArgs    (const char *arg = NULL);
+    void displayArgs   (const char *arg = NULL);
 
-    bool createOutput   (OUTPUTS driver, const SidTuneInfo *tuneInfo);
-    bool createSidEmu   (SIDEMUS emu);
-    void displayError   (const char *error);
-    void displayError   (unsigned int num) { ::displayError (m_name, num); }
-    void decodeKeys     (void);
-    void updateDisplay  ();
-    void emuflush       (void);
-    void menu           (void);
-    void refreshRegDump ();
+    bool createOutput  (OUTPUTS driver, const SidTuneInfo *tuneInfo);
+    bool createSidEmu  (SIDEMUS emu);
+    void displayError  (const char *error);
+    void displayError  (unsigned int num) { ::displayError (m_name, num); }
+    void decodeKeys    (void);
+    void updateDisplay ();
+    void emuflush      (void);
+    void menu          (void);
+    void refreshRegDump();
 
     uint_least32_t getBufSize();
 
@@ -213,11 +213,11 @@ public:
     ConsolePlayer (const char * const name);
     virtual ~ConsolePlayer() {}
 
-    int  args  (int argc, const char *argv[]);
-    bool open  (void);
-    void close (void);
-    bool play  (void);
-    void stop  (void);
+    int  args (int argc, const char *argv[]);
+    bool open (void);
+    void close(void);
+    bool play (void);
+    void stop (void);
 
     player_state_t state (void) const { return m_state; }
 };
